latest update to posdata conversion, fixed bytetrack wrong rounding.

+++ backend/src/backend/backend/models.py
@@ -154,10 +154,11 @@ def delete_file(sender, instance, **kwargs):
         os.remove(path)
         logger.info(f'Removed file with id: {instance.id.hex}')
     if sender == TrackingData:
-        meta_path = media_path_to_file(instance.meta_file.hex, instance.meta_ext)
-        if os.path.exists(meta_path): 
-            os.remove(meta_path)
-            logger.info(f'Removed file with id: {instance.meta_file.hex}')
+        if instance.meta_file:
+            meta_path = media_path_to_file(instance.meta_file.hex, instance.meta_ext)
+            if os.path.exists(meta_path): 
+                os.remove(meta_path)
+                logger.info(f'Removed file with id: {instance.meta_file.hex}')
 
 
 class Plugin(models.Model):
